+++ internal/modules/loanproduct/dto.go
@@ -29,9 +29,9 @@ type ApplyForLoanResponse struct {
 }
 
 type ManualRepaymentRequest struct {
-	LoanID         string `json:"loan_id" binding:"required"`
-	Amount         int64  `json:"amount" binding:"required"`
-	TransactionPin string `json:"transaction_pin" binding:"required"`
+	LoanID         string  `json:"loan_id" binding:"required"`
+	Amount         float64 `json:"amount" binding:"required"`
+	TransactionPin string  `json:"transaction_pin" binding:"required"`
 }
 
 type LoanRepaymentResponse struct {
@@ -51,8 +51,8 @@ type ActiveLoansResponse struct {
 }
 
 type RepaymentRequest struct {
-	Amount      int64  `json:"amount"`
-	RepaymentID string `json:"repayment_id"`
+	Amount      float64 `json:"amount"`
+	RepaymentID string  `json:"repayment_id"`
 }
 
 type LoanDetailsResponse struct {
